Greatly improve startup performance
Multithread the annotation scanning to reduce scan time by half.

Cache the result of that scan so we don't have to rescan every time. This significantly reduces the time spent loading after the first load has happened

Give the Knot ClassLoader to the annotation scanner and filter the scan to not occur in game library packages - no significant change to performance noticed