[1.17.1-] Fix Guava version, drop Forge support
Guice 6.0.0 needs Guava 31, Minecraft 1.17.1 uses an older Guava version. Fabric can bundle the newer version of Guava and resolve the dependency correctly. Forge's JiJ system does not actually support 1.17.1, and Shadowing this Guava version will conflict with the base game and crash, forcing us to either downgrade the project to 5.1.0 and make porting a pain, or remain on 6.0.0+ and drop Forge for these versions.